Doc fix, type fix and option variables fixes

+++ housecimis.c
@@ -115,7 +115,7 @@ static const char *housecimis_status (const char *method, const char *uri,
     int root = echttp_json_add_object (context, 0, 0);
     echttp_json_add_string (context, root, "host", host);
     echttp_json_add_string (context, root, "proxy", houseportal_server());
-    echttp_json_add_integer (context, root, "timestamp", (long)time(0));
+    echttp_json_add_integer (context, root, "timestamp", (long long)time(0));
     int top = echttp_json_add_object (context, root, "waterindex");
     int container = echttp_json_add_object (context, top, "status");
     echttp_json_add_string (context, container, "name", "cimis");
